Draw attention to possible bug found by mypy.
iapws/iapws97.py:4520: error: Unsupported target for indexed assignment ("Tuple[float, float]")
iapws/iapws97.py:4522: error: Unsupported target for indexed assignment ("Tuple[float, float]")

Because tuples are not modifable, the previously code generated mypy
errors despite being what I thought was the intended API.  I thought
that I could modify the code to remove the stealthy update of the
CALLERS par list, but it trurns out that this change is required, and
causes tests to fail.  Was this update intentional?  If so, perhaps a
comment is appropriate?  If not, I think this is a bug.