Convert R, Tref, and Rhoref to lists of length 1.
This may be a controversial change, but for mypy to make much sense of
the static types, we need _constants to be more stongly typed (not
dependent on the key string).  Perhaps there's something I can do to
restore this later, but not right now wre' still getting static types.