Significant mypy additions.
This should be entirely type hints, and some explicitly conversions
from integer to float by adding ".0".  This commit has introduced
nearly 200 mypy type errors, some of which are going to require more
significant thought, so this seemed like a good place to commit even
though it's a bit of a mess at the moment.