Clear up some typing issues surrounding _Pv, _rhoL and _rhoG.
Mostly these changes were about making it easier for mypy to detect
the types of the float lists, but it also made the relatonship between
the base class and the derived classes a lot clearer.  I think the
design would be cleaner with a base class implementation that took
constants as parameters from short stubs in the derived classes.