Better consistency for _visco() and _thermo().
This is first commit that has really changed tests.py.  I don't
understand why all of the other classes would implement _visco() as an
object method, while Air implemented it as a static method.  Air even
went to the trouble of accepting an optional fase parameter, that
wasn't actually required because the _visco static method had no
relationship to the unimplemented object method of the same name?

The cleanest fix in my opinion is to be more explicit that derived
classes are expected to implement _visco() and _thermo(), have Air do
so properly as a class derived from MEoS.  The change to the tests was
minor, the method is ostensibly part of the private API, and it's
consistent with the _thermo() tests immediately below.