More comprehensively cast from numpy.Float64 to Python float.
Since fsolve, newton, and minimize are the primary (only?) internal
sources of the numpy types, I think this means we're really and truly
using Python floats throughout the code now.  That means that I should
be able to eliminate numpy type confusion as the causes of the remaing
mypy errors.