This commit demonstrates that numpy.Float64 is mostly gone.
It's not that I dislike numpy.Float64 some much as there is no good
static typing information for the numpy module, so it's hard to tell
if numpy code does what it seems to. :-( Plus now when we say a method
returns a float, we really mean that, and not just a somewhat
float-like Python object backed by a C extension.