Work around a mypy quirk with type ordering and function definiton.
The problem seems to be that the functions don't get defined in the
order that they appear in the source code, and mypy gets confused
about whether T, P, rho, and A are None or not.  The assertions in
this awkward locatioin make mypy happy.  We could pass the parameters
into fsolve, or perhaps it would be easier to just set all three
values to valid floats in calculable before entering this function.