Improve static typing with a custom class instead of a dictionary.
It's much more difficult to have a static assurances of the keys and
values of a dictionary than a custom class.  The resulting variable
names are slightly different, but not obviously worse to me.  More
importantly, the type is differentiated from alll other dictionaries
that are keyed by str and have float values, of which there are
several, including some with overlapping keys (e.g. _fav).