Split up the thousand line calculo function.
The size of this function was one of the primary reasons that I
started making changes to IAPWS.  While I understood the overall
structure of this function, I was unable to understand what the inputs
and output of each calculation "mode" was because everything was in a
single scope with processing before and processing after the "switch".
This version is far from perfect, and is about 150 lines larger than
the previous version, but I think it's a lot more readable, and it's
likely that the size will return to something more like the orginal
after some additional cleanup, but this seemed like a good place to
commit.