Remove to two unused variables.
While there was a lot of farily arbitrary restructuring in the
previous commit, this change was the one that seemed to be the most
substantively different.  These two variables really do seem to have
just been unused, whereas most of the other changes were more
obviously stylistic or caused by changes in data flow.