Call calculo to determine if state is calculable.
This is beneficial for mypy (and also for humans to some extent)
because it struggles with code paradigms like _mode where variables
are tested for types in one place, then summarized in _mode, and then
used logically in distant code that presumes the logic is sound.  By
placing the actual tests closer to the usage, mypy is able to figure
out, and humans can more easily verify that the constraints are
enforced.