Update IAPWS97 to better match IAPWS95 code structure.
In comparison to the IAPWS95 changes, these changes were hardly needed
because the code was already so much clearer.  There were a couple of
places where mypy had to be "accommodated" to remove the "type:
noginore" comments.  The first was moving the "is not None" tests
closer the calls that used the variables, and the second was splitting
out the _solve_X_X() functions because mypy doesn't cope well with
sub-function declarations and conditional logic.  Both changes were
beneficial anyway in my opinion, drawing attention to an unusual
computation of sigma in two cases, and generally providing strong
assurances that the entire _solve_X_X() call chain doesn't otherwise
update the object.  I wish we had that feature in IAPWS95.